update publish script
- top_level.txt will now be included in the wheel dist-info for
  compatibility with python-vendorize
- bump patch

+++ publish.py
@@ -2,7 +2,10 @@
 # Imports #
 # ------- #
 
+from glob import glob
+from os import path
 from textwrap import dedent
+from zipfile import ZipFile
 import is_git_repo_clean
 import subprocess
 import sys
@@ -36,7 +39,10 @@ if not is_git_repo_clean.checkSync():
 with open("pyproject.toml", "r") as f:
     pyproject = toml.load(f)
 
-    githubUrl = pyproject["tool"]["poetry"]["repository"]
+    poetryConfig = pyproject["tool"]["poetry"]
+    githubUrl = poetryConfig["repository"]
+    moduleName = poetryConfig["name"]
+    version = poetryConfig["version"]
 
 pypiWarning = dedent(
     f"""\
@@ -52,9 +58,20 @@ with open("./README.md", "r") as f:
 with open("./README.md", "w") as f:
     f.write(commentedWarn.replace(pypiWarnComment, pypiWarning, 1))
 
-subprocess.run("poetry publish --build".split(" "))
+subprocess.run("poetry build".split(" "))
+
+pathToTopLevel = f"{moduleName}/top_level.txt"
+if not path.isfile(pathToTopLevel):
+    raise Exception(f"top_level.txt must exist at '{pathToTopLevel}'")
+
+topLevelDistInfo = f"{moduleName}-{version}.dist-info/top_level.txt"
+pathToWheelZip = glob(f"dist/{moduleName}-{version}-*.whl")[0]
+with ZipFile(pathToWheelZip, "a") as wheelZip:
+    wheelZip.write(pathToTopLevel, topLevelDistInfo)
 
 with open("README.md", "w") as f:
     f.write(commentedWarn)
+
+subprocess.run(["poetry", "publish"])
 
 print("donezo!")
